+++ components/app.js
@@ -3,9 +3,9 @@ var express = require('express');
 var path = require('path');
 var cookieParser = require('cookie-parser');
 var logger = require('morgan');
-var router = require('./routes/index');
-var about = require('./routes/about');
-var api = require('./routes/api/index');
+var router = require('../routes/index');
+var about = require('../routes/about');
+var api = require('../routes/api/index');
 
 var app = express();
